Backend/requirements.txt
# FastAPI Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Base de Datos
sqlalchemy>=2.0.0
pymysql>=1.1.0

# Autenticación y Seguridad
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
argon2-cffi>=23.1.0
google-auth>=2.25.0

# Validación
pydantic[email]>=2.5.0
email-validator>=2.1.0

# Variables de Entorno
python-dotenv>=1.0.0

# Multipart (para formularios)
python-multipart>=0.0.6
fastapi-mail>=1.4.1
pydantic-settings>=2.0.0
